Add a shutdown hook to close cleanly
