+++ src/test/java/net/parttimepolymath/kafkatoy/ProducerFactoryTest.java
@@ -1,0 +1,13 @@
+package net.parttimepolymath.kafkatoy;
+
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class ProducerFactoryTest {
+
+    @Test
+    void testMakeProducer() {
+        assertNotNull(ProducerFactory.make("localhost:9091"));
+    }
+}
